feat(desktop): park v1 terminals on detach + re-enable WebGL
v1's `detachFromContainer` was calling `wrapper.remove()`, fully
detaching the xterm canvas from the document on every tab switch.
Switch to v2's parking-container model: append the wrapper into a
hidden body-level container so xterm stays in the document tree
across React mount/unmount cycles. This converges v1 with v2's
lifecycle and avoids GPU compositor churn from repeated detach.

The shared `getTerminalParkingContainer()` helper is lifted from
v2's `terminal-runtime.ts` into `lib/terminal/terminal-parking.ts`
so both surfaces use a single hidden DOM node.

With parking in place, re-enable the WebGL renderer in v1 (reverting
the workaround in #3924). Pattern matches v2's `terminal-addons.ts`:
defer init to rAF, dispose + DOM fallback on `onContextLoss`, and
share `suggestedRendererType` so a failure stops retrying.
