+++ apps/desktop/src/renderer/lib/terminal/terminal-runtime.ts
@@ -7,6 +7,7 @@ import { DEFAULT_TERMINAL_SCROLLBACK } from "shared/constants";
 import type { TerminalAppearance } from "./appearance";
 import { loadAddons } from "./terminal-addons";
 import { installTerminalKeyEventHandler } from "./terminal-key-event-handler";
+import { getTerminalParkingContainer } from "./terminal-parking";
 
 const SERIALIZE_SCROLLBACK = 1000;
 const STORAGE_KEY_PREFIX = "terminal-buffer:";
@@ -116,34 +117,6 @@ function clearPersistedDimensions(terminalId: string) {
 function hostIsVisible(container: HTMLDivElement | null): boolean {
 	if (!container) return false;
 	return container.clientWidth > 0 && container.clientHeight > 0;
-}
-
-// Body-level hidden container that owns wrapper divs of terminals whose
-// React component is currently unmounted (e.g. workspace switch). Keeps
-// xterm attached to the document so it survives provider remounts without
-// a detach/reattach flash — VSCode's setVisible(false) model. Looked up
-// by DOM id so it's HMR-safe (module-level `let` would leak on re-eval).
-// `inert` removes the whole subtree from the tab order and the accessibility
-// tree, and also moves focus out of it — so a parked terminal's internal
-// <textarea> can't receive keystrokes meant for the active pane.
-const PARKING_CONTAINER_ID = "v2-terminal-parking";
-function getParkingContainer(): HTMLDivElement {
-	const existing = document.getElementById(PARKING_CONTAINER_ID);
-	if (existing) return existing as HTMLDivElement;
-
-	const el = document.createElement("div");
-	el.id = PARKING_CONTAINER_ID;
-	el.setAttribute("inert", "");
-	el.setAttribute("aria-hidden", "true");
-	el.style.position = "fixed";
-	el.style.left = "-9999px";
-	el.style.top = "-9999px";
-	el.style.width = "100vw";
-	el.style.height = "100vh";
-	el.style.overflow = "hidden";
-	el.style.pointerEvents = "none";
-	document.body.appendChild(el);
-	return el;
 }
 
 function measureAndResize(runtime: TerminalRuntime): boolean {
@@ -299,8 +272,8 @@ export function detachFromContainer(runtime: TerminalRuntime) {
 	runtime.resizeObserver?.disconnect();
 	runtime.resizeObserver = null;
 	// Park instead of .remove() so xterm survives the React unmount —
-	// see getParkingContainer.
-	getParkingContainer().appendChild(runtime.wrapper);
+	// see getTerminalParkingContainer.
+	getTerminalParkingContainer().appendChild(runtime.wrapper);
 	runtime.container = null;
 }
